+++ src/main/java/ru/yandex/practicum/filmorate/storage/LikeStorage.java
@@ -5,9 +5,9 @@ import ru.yandex.practicum.filmorate.model.Film;
 import java.util.List;
 
 public interface LikeStorage {
-    void addLike(int filmId, int userId);
+    void addLike(int filmId, Long userId);
 
-    void removeLike(int filmId, int userId);
+    void removeLike(int filmId, Long userId);
 
     List<Film> getPopular(Integer count);
 }
